Minor fix in FramePyramid
Summary:
In case a `FramePyramid` was not owning the memory, the pyramid's `memory_` member variable could be empty.
This is fine in general, however, the class has several asserts ensuring that accessed memory is always within valid bounds.
Thus, ensuring that the `Memory` object is aware of the valid memory.

Reviewed By: enpe

Differential Revision: D55039718

fbshipit-source-id: b2fca36d7c3386a11d9414877cb29a4789d4bf34