Removed the concept of valid layers in CV::FramePyramid
Summary:
The `CV::FramePyramid` had layers, and an extra parameter to specify the number of valid layers.
It was intended to narrow an algorithm to a specific number of pyramid layers without significant compute.
However, this also introduced some friction - so that removing this functionality seems to be the best option.

Instead added the possibility to reduce the number of used pyramid layers via `FramePyramid::reduceLayers()`.

Reviewed By: enpe

Differential Revision: D55539298

fbshipit-source-id: 493e1843877d1d5b6c37660537258c5bab4abe11