Add CMake-function to determine the installation directory for libraries
Summary: When building Ocean for certain platforms, multiple build configurations may be required to live alongside each other. For that reason, we have a option to use "long names" for the library installation path. To better re-use the code that defines this path, it is moved to its own CMake-function.

Reviewed By: ASchneiderMeta

Differential Revision: D55609698

fbshipit-source-id: cc37230a374914e922e9d732768b421ff2052a8e