frame handling
Summary:
**Context for this Diff stack**

With the rollout of CMMv2, SensorDataProvider is deprecated because the concept of 'frame types' no longer exists. Instead we have the concept of 'purposes', which are topic streams: 'iot', 'hand', 'passthrough' etc.

There is a new API to access sensor data that is better aligned with CMMv2: sensoraccess.

In order to debug the new API and accelerate it's adoption I need a POC. Due to it's simplicity I'm picking the Ocean Sensors app and the underlying FrameProvider.

**This Diff**

I'm implementing the `latestFrames` logic for the purpose-based streams. The logic is significantly simpler because we just forward the request to the frame set consumer. There is no need to combine multiple streams into a single one, because the 'active_slots' purpose already does that.

========

Differential Revision:
D54939409

Privacy Context Container: L1124100

fbshipit-source-id: 83477c353c158948a52c62e12ed56f0c9835a75c