allow to process incomplete frame sets
Summary:
**Context for this Diff stack**

With the rollout of CMMv2, SensorDataProvider is deprecated because the concept of 'frame types' no longer exists. Instead we have the concept of 'purposes', which are topic streams: 'iot', 'hand', 'passthrough' etc.

There is a new API to access sensor data that is better aligned with CMMv2: sensoraccess.

In order to debug the new API and accelerate it's adoption I need a POC. Due to it's simplicity I'm picking the Ocean Sensors app and the underlying FrameProvider.

**This Diff**

Incomplete frame sets need to be explicitly supported, especially in case of face/eye cameras. With CMMv2 it is expected that many frame sets contain only the eye tracking cameras.

========

Differential Revision:
D55469059

Privacy Context Container: L1124100

fbshipit-source-id: 51b1c78ad875ef33ccac5b9c9f165e23a53a26c5