implement purpose-based frame processing
Summary:
**Context for this Diff stack**

With the rollout of CMMv2, SensorDataProvider is deprecated because the concept of 'frame types' no longer exists. Instead we have the concept of 'purposes', which are topic streams: 'iot', 'hand', 'passthrough' etc.

There is a new API to access sensor data that is better aligned with CMMv2: sensoraccess.

In order to debug the new API and accelerate it's adoption I need a POC. Due to it's simplicity I'm picking the Ocean Sensors app and the underlying FrameProvider.

**This Diff**

I'm implementing the logic to register, start and stop purpose-based streams. The semantics are slightly different because streams need to be created once and attached to the dispatch thread before starting any data streams. Afterwards they can be started/stopped at will without the need to reinitialize anything.

========

Differential Revision:
D54906814

Privacy Context Container: L1124100

fbshipit-source-id: 4d3e3422f7832feca02e654f9871622035ac9cd4