allow multiple reconfigurations
Summary: CMMv2 may invoke the configuration sequence multiple times, so we need to prepare for that. The canonical solution is to allocate the array in `setNumCameras()`, and then just update the corresponding position in the `setCameraConfiguration()` call, under mutex protection like is already the case.

Differential Revision:
D55753193

Privacy Context Container: L1124100

fbshipit-source-id: cabf9276498dd17d2a6d2ea38c5fc72b219f26eb