initial boilerplate for appsensoraccess
Summary:
**Context for this Diff stack**

With the rollout of CMMv2, SensorDataProvider is deprecated because the concept of 'frame types' no longer exists. Instead we have the concept of 'purposes', which are topic streams: 'iot', 'hand', 'passthrough' etc.

There is a new API to access sensor data that is better aligned with CMMv2: sensoraccess.

In order to debug the new API and accelerate it's adoption I need a POC. Due to it's simplicity I'm picking the Ocean Sensors app and the underlying FrameProvider.

**This Diff**

Add basic boilerplate that opens the library and queries for the available purposes. This is similar to the existing logic that queries for the frame types.

========

Differential Revision:
D54897457

Privacy Context Container: L1124100

fbshipit-source-id: 422e8292b8b209e08c57215652ed0e3fbb69ae01