use v4::FrameSetConsumer internally
Summary:
**Context for this Diff stack**

With the rollout of CMMv2, SensorDataProvider is deprecated because the concept of 'frame types' no longer exists. Instead we have the concept of 'purposes', which are topic streams: 'iot', 'hand', 'passthrough' etc.

There is a new API to access sensor data that is better aligned with CMMv2: sensoraccess.

In order to debug the new API and accelerate it's adoption I need a POC. Due to it's simplicity I'm picking the Ocean Sensors app and the underlying FrameProvider.

**This Diff**

In fbsource we have switched to the v4::FrameSetConsumer and related types, and the new API also exposes these versions. Therefore let's migrate the internal processing to v4 and provide an adapter for v3 since the SensorDataProvider API exposes v3 types.

========

Differential Revision:
D54901995

Privacy Context Container: L1124100

fbshipit-source-id: a83a0ed914f8dc188a25516baaaf35c14343c78e