add exposure settings methods for purpose-based streams
Summary:
**Context for this Diff stack**

With the rollout of CMMv2, SensorDataProvider is deprecated because the concept of 'frame types' no longer exists. Instead we have the concept of 'purposes', which are topic streams: 'iot', 'hand', 'passthrough' etc.

There is a new API to access sensor data that is better aligned with CMMv2: sensoraccess.

In order to debug the new API and accelerate it's adoption I need a POC. Due to it's simplicity I'm picking the Ocean Sensors app and the underlying FrameProvider.

**This Diff**

I'm copying the existing logic to work with stream key. The main change is the different lookup (using the stream key to locate the purpose frame set consumer) and not the frame type.

========

Differential Revision:
D54943821

Privacy Context Container: L1124100

fbshipit-source-id: 97ff511e63d9a887a08f10c22399d433e0d060d5