Improve decoding of QR codes
Summary:
The QR code standard (ISO 18004:2015) defines an optional mirroring feature for QR codes. This diff enables this as a backup mechanism, in case the initial decoding (of the "un-mirrored" code) has failed.

This is now necessary as there are online code generators that use this, for example

https://qrfy.com/

Without this mirroring, codes generated on that site will fail to decode (detection itself is not a problem)

Differential Revision: D56003996

fbshipit-source-id: 4190e792a3a7798ac4a80d8b7a6a1f62d52897e4