Added two new FrameConverter::Comfort::convert() functions
Summary:
The two new functions simplify the conversion of frames.

The dimension of the target frame type was never used (it had to be identical to the source frame).

Thus, now adding two functions replacing the `const FrameType& targetType`:
1) Specify the target pixel format and target pixel origin
2) Specify the target pixel format

Marked other functions as deprecated.

Further, updated the existing unit test.

Reviewed By: enpe

Differential Revision: D56279558

fbshipit-source-id: a8256697a87b0379e3e9206819c9b108dcd83e45