Remove calls to std::strchr()
Summary:
Previously, `std::strchr()` was used to determine whether a given character value parameter `ch` matches one of the characters in the string parameter `str` by checking whether the return value is a null pointer.  But `std::strchr()` returns non-null pointer to the terminating null character when `ch` is `\0`.  This was not the desired behavior.  Replacing with code that iterates over character set data.

Also included:
- Minor changes to comments and test code.
- Change type of `NUMERIC_CHARSET ` in ".../testqrcodes/Utilities.cpp" to match that of `QRCodeEncoder::Segment::ALPHANUMERIC_CHARSET`.

Reviewed By: enpe

Differential Revision: D53504597

fbshipit-source-id: e5214ed60ee6fa2a16e0531455c2cd6abb79ec87