Add variable to CMake to configure the third-party location
Summary:
Configuring the location of the appropriate third-party libraries (based on the Android ABI and build type) in Gradle isn't quite straight-forward.

The easy work-around is to have Gradle provide the root location of ALL third-party builds and let CMake choose the appropriate one.

This will be changed once someone figures out how to do this directly in Gradle. Until this this stop-gap solution will do.

Differential Revision: D56382599

fbshipit-source-id: b1409c2421769084f57937795fd3bb63b318ec1a