Disable OpenCV in third-party build script by default
Summary: OpenCV is very large, takes a long time to build, and it's only needed for some side-by-side comparison with Ocean in a single test. For that reason, let's disable it by default for now. We'll still need to add some parameter to enabled it from the command-line, if someone wants to have it.

Differential Revision: D56471563

fbshipit-source-id: 2e9d453d71b548653e13403566548c2cdbffcba0