Added abstraction for test data
Summary:
Some of Ocean's unit tests are based on test data like e.g., images or videos etc.
As the test data is stored in Meta's infrastructure like e.g., Everstore, we need to add a dedicated abstraction for test data so that we can opensource the tests without exposing any internal infra/APIs.

Therefore, we are adding `TestData`, `TestDataCollection`, and `TestDataManager` to `Ocean::Test`

Reviewed By: enpe

Differential Revision: D56428112

fbshipit-source-id: 340e01303b6dec7920aa72fdf0c3cf343ccc3107