Add script to build third-party libraries for iOS
Summary:
Similar to the other scripts in this directory, this script builds the third-party libraries that are required to build Ocean for iOS.

The usage is the same as for the other scripts, just execute it (on macOS).

Also improved the documentation of the Android script, on which this new script is based.

Differential Revision: D56489677

fbshipit-source-id: 9eb8c18688b9222f7c5fd80987e5797fac87b780