Switching to CPM 0.39
Summary:
The CPM version used here is 1.0.0-development-version, which causes some warnings about being unstable. So, let's switch to the latest release, which is 0.39. Interestingly, the only difference between these two versions is the version number, no other code changes landed since that release. It still makes sense to apply this, though, because it maps to a fixed, public commit:

https://github.com/cpm-cmake/CPM.cmake/tree/v0.39.0

Reviewed By: ASchneiderMeta

Differential Revision: D56473810

fbshipit-source-id: 4e76260f8d50b161ee3b12fa0ef45b2b208c3210