Minor fix
Summary:
The register/unregister function must only be run in static builds. When building this library as a dynamic library in CMake, it will cause compiler errors.

Why does this build with Buck, though? Probably because of the generous use of `preferred_linkage = "static",  # workaround to improve dead-code stripping` in all Android targets, which cause them to be linked statically.

Differential Revision: D56546819

fbshipit-source-id: 6e5d4599fd9e9dd7dffe5fd572a1cb2557f33c68