Remove third-party/android from the CMake config
Summary:
One part of the files provided by this package can be found in the NDKs:

```
% find . -type f -iname 'android_native_app_glue*' | sort
./ndk/21.4.7075529/sources/android/native_app_glue/android_native_app_glue.c
./ndk/21.4.7075529/sources/android/native_app_glue/android_native_app_glue.h
./ndk/25.1.8937393/sources/android/native_app_glue/android_native_app_glue.c
./ndk/25.1.8937393/sources/android/native_app_glue/android_native_app_glue.h
./ndk/26.2.11394342/sources/android/native_app_glue/android_native_app_glue.c
./ndk/26.2.11394342/sources/android/native_app_glue/android_native_app_glue.h
```

and the header file ARCore, `arcore_c_api.h`, will be provided in a dedicated package that also downloads the required corresponding `.so` files (cf. previous diff).

Because of that, it's probably best to remove this package for now.

Differential Revision: D56642959

fbshipit-source-id: 068983d77cd735952d693772ba2b530673cc3497