Add a CMake config for ocean_rendering_glescenegraph_qt
Summary:
Qt is currently not part of the automated third-party builds for Ocean (and it's not planned to add it). So, this target will be optional, i.e., it's defined when Qt is found otherwise it isn't.

Regardless, this will have to be tested more, which will happen in a separate step later.

Differential Revision: D56780727

fbshipit-source-id: 6b732415704ba10436fd931c1e20bde5cf69125c