Add CMake config to find HarfBuzz lib
Summary:
This library is used by freetype. Without explicitly adding this library to Ocean, builds will fail, for example on Android it's either because linking fails for static builds or because the `.so` isn't packaged into the  APK)

Adding this file will help to make sure the library is found.

Differential Revision: D56951352

fbshipit-source-id: 6a320c9c171f4a07322b960c9546fbb1bf686962