Fix inifnite recursion in requestPermissions()
Summary:
Requesting the same permission multiple times before a response is received is causing an infinite recursion in requestPermissions().
This change protects against this in two ways: (1) Handle an empty response correctly in onRequestPermissionsResult(), and (2) prevent duplication of pending permission requests in requestPermission().

The issue was detected in the UVTextureMapping demo app. Without this change, the app crashes with a stack overflow. With this change, the app launches succesfully and prompts for camera permission. Note that after granting permission, the app must be restarted before it can successfully access the camera and this issue is being tracked separately (T183039266).

Reviewed By: enpe

Differential Revision:
D57021477

Privacy Context Container: L1191897

fbshipit-source-id: 17714f7c8c85514da1d1dea4afa7c77684a7dbcb