Add impl/metaonly/ocean to the Doxygen config
Summary:
Adding `impl/metaonly/ocean` to the Doxygen config. Internally, this will work as expected. Externally, i.e. on Github, there will be a warning in the logs similar to this one:

```
warning: tag INPUT: input source 'impl/intentionallydoesntexist' does not exist
warning: source 'impl/intentionallydoesntexist' is not a readable file or directory... skipping.
```

But otherwise this will be ignored.

Reviewed By: janherling

Differential Revision: D57642692

fbshipit-source-id: 6b7244605d36137d95ff9d1c722cd242a2b586d0