Rename pull request template
Summary:
Renaming it because the automated repo checkup doesn't seem to check this location correctly.

According to the GH docs at

https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository

it is allowed to have it at

```
.github/PULL_REQUEST_TEMPLATE/pull_request_template.md
```

But according to

https://www.internalfb.com/intern/opensource/github/repo/1133636141337819/checkup/

it only seems to check for files, not directories:

```
Error: not found: (PULL_REQUEST_TEMPLATE*, .github/PULL_REQUEST_TEMPLATE*)
```

So, let's test that hypothesis.

Reviewed By: janherling

Differential Revision: D57700113

fbshipit-source-id: f6b41eeb7a995fdab3852199f5917a10735a173f