Split source file
Summary:
There is build error on macOS when build Objective-C code:

https://fburl.com/phabricator/dbf15tf8

This is likely caused by the file type, `.cpp`, for which the compiler is set to build C++, not Objective-C.

To fix this, the file `ocean/io/Signature.cpp` will be split into a `*.cpp` and a `.mm` part.

This problem only appears in the CMake version of the Ocean builds because that's where OpenSSL is available.

Also, the new `Signature.mm` is **not** added to the XCode project because `Signature.cpp` isn't included either (because of missing OpenSSL support for the XCode project)

Reviewed By: janherling

Differential Revision: D57931960

fbshipit-source-id: b0904d97d59157780006cde6d4200ea09c5821ba