Minor fix in CMake config of ocean_platform_wxwidgets
Summary:
When building Ocean OSS with CMake on a devserver, there is an error about `wxWidgets::wxWidgets` not existing (full error: P1385159208).

According to

https://cmake.org/cmake/help/latest/module/FindwxWidgets.html

that target name seems to be introduced with CMake 3.27.

The devservers are using CMake 3.26, which explains the error.

Using `wxWidgets_LIBRARIES` and `wxWidgets_INCLUDE_DIRS` instead fixes the issue.

Reviewed By: janherling

Differential Revision: D57994347

fbshipit-source-id: 20b783ebef314746e63fddb9e8cca272b9223964