Split DEVELOPMENT_TEAM from XCode projects
Summary:
With this change it will be possible to completely remove any reference to Apple Team IDs or provisioning profiles from the project files. To use this, do the following

```
cp development_team.xcconfig.template development_team.xcconfig
```

And then define the variables inside the cloned file. Because `development_team.xcconfig` is listed in `.gitignore` it will not be committed.

For now, this is only converting one demo to this new way of defining the code signing credentials, `democvdetectorlinedetectorios`. Adding the remaining demos should be straight-forward.

Reviewed By: janherling

Differential Revision: D58263193

fbshipit-source-id: a92d443f06106678094b3b71514b69e427bc3163