Add platform-dependent define for wxwidget handles
Summary:
The library `ocean_platform_wxwidgets` was never used on Linux before. So, the type `WXHWND` isn't known and causes build errors. For Linux, `WXWidget` must be used.

This diff adds a macro that is defined depending on the platform that the code is built for.

Note: generally, neither `WXHWND` nor `WXWidget` should be used directly (because they are platform-dependent). However, this is a pre-existing problem that will have to be addressed separately. Here, we're focussing on making the code compile.

Reviewed By: janherling

Differential Revision: D58148554

fbshipit-source-id: e6fcfea5519b36c2bc1359efa87792fcf912ae16