Removing team ID and provisioning profiles from testopencv
Summary:
This project is currently broken. It's not worth spending much time on this right now, so only the ID of the development team and the provisioning profiles will be removed.

Fixing this project and adding the `deveopment_team.xcconfig` can be done at a later time should this project ever become of interest.

Reviewed By: janherling

Differential Revision: D58432457

fbshipit-source-id: cb1b763d2236f1b7133bc8d50e96040bc64777e6