ocean | Don't lock on destruction of WorkerPool.
Summary:
Fixing lock order inversion, and given that this is explicitly in the destructor - locking on free/used is a bit weird, since it only could happen if someone randomly would walk in async and try to add or remove a given worker.

Lock order is here:
- Processor::get() here https://fburl.com/code/scapmto6, which accesses the OceanManager::get() to register the singleton
  - WorkerPool::lock_
  - OceanManager::registerSingleton - OceanManager::lock_
- OceanManager destructor here, goes inverse:
  - Lock OceanManager::shutdown() - OceanManager::lock_
  - ~WorkerPool - WorkerPool::lock_

This is primarily fixing a thread sanitizer problem only, since that's what caught the lock order inversion.
This doesn't seem to affect prod in any way, since would be triggered on exit, and even if it does - the process was already exiting anywho.

Reviewed By: janherling, francoiscoulombe

Differential Revision: D58572575

fbshipit-source-id: 084a39072a6bd138a1e3a5a8411a63bc11934138