Disable Linux-build of ocean_platform_wxwidgets in CMake
Summary:
It turns out that the code of `ocean_platform_wxwidgets` is not platform-independent enough to work with wxwidget that uses GTK3. This issue will have to be addressed separately.

In order to enable builds in the Github CI, `ocean_platform_wxwidgets` will be disabled on Linux until the above problem has been addressed.

Reviewed By: janherling

Differential Revision: D58637867

fbshipit-source-id: a005e627d136323a696b558446e0bdeefd83085d