Minor changes to Github workflow file
Summary:
Make it so that the installation of all apt packages are auto-confirmed (`-y`).

Added the possibility to install a newer version of CMake. This is disabled by default as it is only required for debugging when using `act`. The Docker image of `act` currently don't have Ubuntu 24.04 yet, only Ubuntu 22.04.  The CMake version, 3.22, in that image is too low for the current CMake config of Ocean (must be >= 3.26).

Reviewed By: janherling

Differential Revision: D58637865

fbshipit-source-id: c6e4ac086ecb595174a060d5c35596e111ab6f4f