Disable some external codecs in libtiff in CMake
Summary:
For a, as of yet, unknown reason, there are some difference between the Linux builds on Fedora (company laptop) and Ubuntu (Github and Docker on company laptop). On the Fedora build, some external codecs in libtiff are disabled by default while the same codecs are enabled by default on Ubuntu. That causes the Ubuntu build to fail.

Additionally, there are problems finding some of the third-party dependencies of libtiff in the Ubuntu build. For example, [`FindLibLZMA.cmake`](https://cmake.org/cmake/help/v3.29/module/FindLibLZMA.html) provide the target `LibLZMA::LibLZMA` on the Ocean side (this is the CMake default name). In libtiff, however, it's called [`Findliblzma.cmake`](https://github.com/libsdl-org/libtiff/blob/master/cmake/Findliblzma.cmake) and provides the target `liblzma::liblzma` - this causes problems on a case-sensitive OS. It's unclear why this work on Fedora, though.

In either case, the external cases which cause problems will be explicitly disabled now.

Reviewed By: janherling

Differential Revision: D58637866

fbshipit-source-id: f9eae0e26effeb921a23379c5c90bb9ad317c10e