Add workflow to build the Github website
Summary:
This new workflow builds and deploys the website of Ocean. However, deployment is currently disabled, so it will locally build the website but not publish it (sort of a dry-run). Deployment will be enabled once the repository is going public.

This workflow is configured to rebuild the website (+ the Ocean doxygen pages) every time something in `ocean/website/*` changes. It will also rebuild everything every 24h (to reflect changes to the doxygen changes).

Reviewed By: janherling

Differential Revision: D58638549

fbshipit-source-id: 8e56dc1784d6d48ac8797fb98434027ca685a896