Disable CMake target ocean_devices_arkit_roomplan_swift
Summary:
The Swift bridge headers required for `ocean_devices_arkit_roomplan_swift` are not created. As a result, the Swift compiler complains about missing symbols:

https://fburl.com/phabricator/yrnvll8b

Since this target isn't currently used anywhere else, it will be disabled for now to unblock the other iOS builds. A fix will be investigated separately (tracking this in T193047951).

Reviewed By: janherling

Differential Revision: D58792009

fbshipit-source-id: eac26540db7ed16a7a4bf5f34da1b6f15d03e624