Added Frame::updateMemory()
Summary:
The new function(s) allow to update/change the externally owned memory of a frame.
This function can be helpful in case an image source is constantly delivering new frame data (located in an external memory block), and the same `Frame` object is intended to use the external memory.

Added a corresponding unit test.

Reviewed By: enpe

Differential Revision: D58893380

fbshipit-source-id: 7915c80c8785f9c904067facada908d53dafb5e1