Split testcv job in github workflow
Summary:
The Debug testcv job is by far the longest-running job in the build workflow. This change divides it into multiple jobs for better parallelization.
The subsets were chosen by looking at runtimes reported on github as well as aggregate times when running the test subsets on a local machine (see P1475567941 for details).
Also move an unrelated test exclusion that happened to be in the wrong job.

Reviewed By: enpe

Differential Revision:
D59602125

Privacy Context Container: L1191897

fbshipit-source-id: 1d72cfd45fcc1de47d91daaaeb1161db1dbb424b