[third-party] Disable additional dependencies of libtiff
Some the dependencies of libtiff cause build errors, cf. issue #9 and potentially #15.

Test plan:

Confirm that the build of the third-party libraries and Ocean no longer fails because of missing `JBIG::JBIG` or similar

https://github.com/facebookresearch/ocean/blob/main/building_for_macos.md
https://github.com/facebookresearch/ocean/blob/main/building_for_ios.md

Note that there are still some other build failures, which will be addressed in separate diffs.