[third-party] Disable harfbuzz
On Macs (Intel-based), when `libharfbuzz*.a` exists in a system-level directory, CMake currently selects that version and not the version that comes with Ocean.

This causes iOS builds to fail because CMake doesn't seems to notice that's intended for a different platform.

To work-around this problem, this diff disables the build and use of harfbuzz altogether.

If harfbuzz is not absolutely required by freetype (or Ocean), we should probably remove it completely to reduce build times and complexity.