Remove test jobs
Summary:
Because some tests are flaky and we don't want to conflate test failures with build errors, we are removing test jobs for now, with the expectation that we will add them back as a seperate dependant workflow later.
Archive and upload install directory instead of build directory for now, but when or before we re-enable testing, we will need to re-add the build archives.

Reviewed By: enpe

Differential Revision: D59761283

fbshipit-source-id: 7b54f620623c63dcaf33c25cc9203e5f905cc9fe