Remove pointer overflow in Neon enabled builds of reverseRowPixelOrderInPlace()
Summary:
Fixes pointer overflow in Neon enabled build of `FrameChannels::reverseRowPixelOrderInPlace()` when `size` is less than 16.

Impact of this change is not observable in normal builds.  But this change removes noise and halting behavior in some builds with sanitizer options enabled, making those builds more useful for testing and debugging.

Reviewed By: janherling

Differential Revision: D59763519

fbshipit-source-id: c0e0b0e73f785eb1309d9d84940123ce343be597