Added new demo app ImageReceiver
Summary:
Just a new demo application which allows to receive images sent via network (via Ocean's `PackagedTCPServer` and `PackagedTCPClient`).

The app is connecting to a `PackagedTCPServer` and waiting to receive concoded images. Whenever a new encoded image arrives, the image is displayed.

Currently, this demo is a pure macOS demo.

Reviewed By: ASchneiderMeta

Differential Revision: D60057699

fbshipit-source-id: d97b330e84e9aa45559585304fd9b5ee050cb85e