Extended LiveVideo with modes for exposure, ISO and focus
Summary:
In preparation for additional configuration possibilities of USB cameras, we need to extend `Media::LiveVideo` to allow more control over camera controls like exposure duration, focus, etc.

Therefore, adding the `LiveVideo::ControlMode` enum which allows to determine which control mode is used for individual camera controls.

Further, extended several functions so that the control mode can be queried.
The explicit control mode now also changes the behavior of the return values of functions like `LiveVideo::exposureDurtation()`, etc. so that we remove the default values in this commit - to ensure that we catch every single usage of these functions (so that we can adjust the code logic).
We will bring back the default values in the next commit.

Finally, fixed focus behavior on Android phones (a valid focus value was never returned before).

Reviewed By: ASchneiderMeta

Differential Revision: D59940919

fbshipit-source-id: df6a59988120bc9efb815d422393063ee0febe9a