Extended ExternalCamera demo app for Quest
Summary:
Now, the demo allows to control focus and exposure in case the connected camera supports the feature.

The user can use the left and right joysticks to change the manual mode of focus and exposure, or to activate the auto mode.

In case a mode is not supported, a text output is provided.

Differential Revision: D60794221

fbshipit-source-id: 7dbacc8f538d77d6d0482ed193582a5822154b52