Improved swapchain behavior in GLESceneGraph for Quest
Summary:
When OpenXR fails to wait for the swapchain, we cannot simply release the swapchain and eventually render the next frame.
Instead, we need to keep the swapchain image and we need to try waiting for the image during the next render call.

Further, we need to ensure that we re-start with the same framebuffer(left eye or right eye) for the next render call.
Otherwise, the framebuffers will be swapped (left/right).

Reviewed By: ASchneiderMeta

Differential Revision: D61040110

fbshipit-source-id: 20b70f621f902dd6057314efab51a85a292cf556