Improve computation of the QR code poses
Summary:
To improve the computation of the poses of QR codes, the implementation tries to find a fourth point correspondence (the first three being the finder patterns).

Previously, the first alignment pattern (between the top-left and top-right finder patterns) was selected. This isn't optimal because this pattern is *almost* colinear with these finder patterns, which can lead to numerical problems.

Now, the alignment pattern is chosen based on the version of the QR code
* V1 - has no alignment patterns and is ignored
* V2-V6: only have one pattern, so it is used
* V7-V13: choose the last pattern in the bottom-right corner to avoid potential problems with colinearity
* V14-...: choose the pattern that is one column to the right and one to the bottom.

Reviewed By: leapally

Differential Revision: D61301142

fbshipit-source-id: 9ff42a35795691193e96b1687fe4f25fc9e26e7b