Update the way search window for QR code alignment pattern detection is determined
Summary: Update to determine alignment pattern search window more reliably.  If the top left and bottom right corners of the search window in object space is projected to points in image space that are close to being horizontal or vertical to each other, the resulting search window could be narrow and miss the alignment pattern.

Reviewed By: enpe

Differential Revision: D61298595

fbshipit-source-id: f322ae56b23a17d7ff6d288f4a6f33bee4a5a8ea