Add missing pixel format types
Summary:
FrameType::FORMAT_ABGR32 & FrameType::FORMAT_YUVA32 are missing and we assume that even that are not very common on iOS ecosystem* and thats why we are getting a small number of error cases ([auto crop unidash results](https://fburl.com/unidash/nz5shxi0))

* According to Metamate: Based on the iOS ecosystem, here are the formats sorted by their commonality, with the most common first:
1. FORMAT_RGBA32: 40-50%
2. FORMAT_BGRA32: 30-40%
3. FORMAT_ARGB32: 15-25%
4. **FORMAT_ABGR32**: 5-15%
5. **FORMAT_YUVA32**: 2-5%
6. FORMAT_YA16: <2%

Note: From [unidash](https://fburl.com/unidash/nz5shxi0) for SmartCrop we can see that image sources are not include the missing types and as that image sources data are from finished events it's an evidence that possible the missing formats are the issue. I would be nice if pixel format could logging also on fail events. cc: ctsimpouris

Reviewed By: foivos-acc, janherling, pappach

Differential Revision: D64459039

fbshipit-source-id: 7560a85ed2ddd6cd3bb05e44bae6e40c77a70975