Avoid infinite loop in tests
Summary:
Some NonLinearOoptimization tests generate random points in an infinite loop until a suitable point is found. Most of the time, the first point is sufficient. Some of the time, it may take a few tries or even a few hundred. In very rare cases, it no randomly-generate point is suitable.
Replace infinte loops with loops that terminate when test duration has been exceeded.  If not enough suitable points have been generated, this is a problem with test case creation, not a problem with the optimization, so skip the incomplete case.

Reviewed By: leapally

Differential Revision:
D64439452

Privacy Context Container: L1191897

fbshipit-source-id: 469dce79ec592e56f6f479c1b961e1f476078594