Add ScopedState to Passthrough
Summary:
For app with multiple modules which may use Passthrough (PT) for their visualization, this change allows app to store the state of PT (on/off) before changing it. This helps avoid leaving PT on when the previous had it disabled, and vice versa.

This is similar to the `ScopedState` of the hand visualizer.

Reviewed By: janherling

Differential Revision: D65095168

fbshipit-source-id: 643e01e9a54e13be3fb990ba95d3de61d74b901b