Remove unused-variable in ../xplat/ocean/impl/application/ocean/demo/misc/appinspector/BinaryAnalyzer.cpp +1
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Differential Revision: D68509679

fbshipit-source-id: ea19b6e77b573798bda4250eecc3ccc93cb670c1