Remove unused-variable in ../xplat/ocean/impl/ocean/devices/arkit/AKFactory.mm +1
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D68568574

fbshipit-source-id: e1225217ca44d77267b2a1f341dcf13b0bc79945