Convert buildDate to use std::string_view
std::string_view is the modern C++ replacement for null-terminated char
pointers. In this specific function, it has two advantages:

First, instead of constructing four new strings (dateStr, yearStr,
monthStr, and dayStr), it is possible to use slices that reference parts
of the original string_view instead. This saves four copies.

Second, the initialization of the "months" array can be moved to
compile-time, since std::string_view is constexpr (unlike std::string).

This commit also replaces atoi with from_chars, as atoi requires a
null-terminated string. This is beneficial, because from_chars provides
an error code indicating whether or not the parsing was successful.