Fix QRCodeDetector3D
Summary:
# Problem

QA reported codes spawning in unexpected locations when two codes are in the camera view (cf. attached task). This is what the bug looks like (XRPlayground):

https://pxl.cl/74b3Z

# Analysis
The problem seems be a combination of

1.  incorrectly matching candidates in the left and right images: https://pxl.cl/74R8P
2. only extracting the code in one image (without checking its counterpart is actually the same code).

# Fix
To fix this, the detector will now decode both of the matched candidates in order to enforce that they belong to the same code:

https://pxl.cl/74RCK

Reviewed By: manderyao

Differential Revision: D73608523

fbshipit-source-id: 7a22b541d05b99074cbae37abb21f04945d8e5f6