Fix Cmake minimum version 3.5 conflict in libtiff this time in Windows
Summary:
In a few of ocean's 3P libs (libtiff and vtzero), they explicitly require:
cmake_minimum_version(3.2), which conflicts with CMAKE's most recent deprecation of CMAKE < 3.5. This causes local build to fail.
Here we fixed this by following CMAKE's suggestion to add -DCMAKE_POLICY_VERSION_MINIMUM=3.5 to fix this.

This was fixed previously in unix/mac/android in D73386709 but was not also fixed in windows builds. See that Diff for more info.

Differential Revision: D75242952

fbshipit-source-id: 79e307d9881b66b7b59bf44f369aeed502b1ae9c