Added default build directories to .gitignore
Summary: When running the build scripts for both third-party and ocean without any directories, the directories are added into the ocean directories with fixed names. Since these default build and install directories are not in .gitinclude, both git and sapling/mercurial will track massive numbers of files. This change adds these directories to the .gitignore file so it does not display lots of files in status output.

Differential Revision: D75482587

fbshipit-source-id: a687724b61a74bc5f113632b0822de6bdf6998a9